requirements.txt
# Core libraries
pandas>=1.5.3
numpy>=1.20.0
matplotlib>=3.4.0
networkx>=2.6.0
tqdm>=4.61.0

# Data processing
jsonlines>=2.0.0

# Graph visualization
plotly>=5.3.0
seaborn>=0.11.0
pyvis>=0.3.2

# Streamlit app
streamlit>=1.22.0
pillow>=8.3.0

# RAG components
sentence-transformers>=2.2.0
scikit-learn>=1.0.0

# LangChain (newer structure)
langchain>=0.0.267
langchain_openai>=0.0.2
langchain_core>=0.1.0

# File handling
pathlib>=1.0.0

# Neo4j and graph tools
neo4j>=5.8.0
py2neo

# PDF processing
unstructured_inference
pdf2image
pdfminer.six==20221105
camelot-py[cv]>=0.11
tabula-py
sec_parser

# Other utilities
python-dotenv
openai>=0.28.0
textwrap3>=0.9.2